Relocate type_hinting module
`valid_value` isn't currently used, but I like the test so much and it's quite small, so I'd rather just keep it around
